backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
psycopg==3.1.19
pydantic==2.11.7
pydantic-settings==2.4.0
pydantic_core==2.33.2
pytest==8.3.2
pytest-mock==3.14.0
sniffio==1.3.1
sqlalchemy==2.0.31
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
alembic==1.13.2
httpx==0.27.0
PyYAML==6.0.1
pytest-asyncio==0.23.7
